Add docker-compose.yml and update Dockerfiles
A new docker-compose.yml configuration file was created to define the system architecture, which includes 'resource-postgres', 'song-postgres', 'resource-service' and 'song-service'. The Dockerfiles in both the 'resource-service' and 'song-service' directories have also been extensively updated, providing detailed steps for building the application, including instructions for preparing dependencies and packaging the application.