Enable Eureka Discovery Client in ResourceServiceApplication
The ResourceServiceApplication is now a Eureka Discovery Client. These changes involve the addition of Spring Cloud and Eureka Client dependencies in the pom.xml file, enabling the Discovery Client in the main application file, and an update to the application.properties file to specify the Eureka Server URL. The database URL has also been adjusted.